pyproject.toml
[project]
name = "xplore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
